requirements.txt
aiogram
python-dotenv
SQLAlchemy
sqlmodel
streamlit
yookassa
